+++ packages/at_client/example/bin/collections_todos.dart
@@ -196,7 +196,7 @@ class _TodosAppState extends State<TodosApp> {
     collection = await atClient.collection<Todo>(
       'todos.$ns',
       const Duration(days: 365),
-      eventsFromLocalSecondary: true,
+      eventSource: EventSource.both,
       fromJson: Todo.fromJson,
       typeTag: 'Todo',
       cleanupOrphansOnCreation: true,